+++ ContentPatcher/Framework/Migrations/Migration_1_25.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics.CodeAnalysis;
+using StardewModdingAPI;
+
+namespace ContentPatcher.Framework.Migrations
+{
+    /// <summary>Validates patches for format version 1.25.</summary>
+    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Named for clarity.")]
+    internal class Migration_1_25 : BaseMigration
+    {
+        /*********
+        ** Public methods
+        *********/
+        /// <summary>Construct an instance.</summary>
+        public Migration_1_25()
+            : base(new SemanticVersion(1, 25, 0)) { }
+    }
+}