+++ ContentPatcher/ModEntry.cs
@@ -54,7 +54,8 @@ namespace ContentPatcher
             new Migration_1_21(),
             new Migration_1_22(),
             new Migration_1_23(),
-            new Migration_1_24()
+            new Migration_1_24(),
+            new Migration_1_25()
         };
 
         /// <summary>The special validation logic to apply to assets affected by patches.</summary>